fix(nextjs): Parse StringOrURLFnToString correctly in authMiddleware
