+++ .changeset/wise-rockets-fetch.md
@@ -1,0 +1,5 @@
+---
+'@clerk/nextjs': patch
+---
+
+`authMiddleware` handles multi-domain relared properties passed as functions properly
