+++ packages/nextjs/src/server/authenticateRequest.ts
@@ -11,13 +11,13 @@ import {
   JS_VERSION,
   PUBLISHABLE_KEY,
   SECRET_KEY,
-  SIGN_IN_URL,
 } from './clerkClient';
 import type { WithAuthOptions } from './types';
-import { apiEndpointUnauthorizedNextResponse, getCookie } from './utils';
+import { apiEndpointUnauthorizedNextResponse, getCookie, handleMultiDomainAndProxy } from './utils';
 import { decorateResponseWithObservabilityHeaders } from './withClerkMiddleware';
 
 export const authenticateRequest = async (req: NextRequest, opts: WithAuthOptions) => {
+  const { isSatellite, domain, signInUrl, proxyUrl } = handleMultiDomainAndProxy(req, opts);
   const cookieToken = getCookie(req, constants.Cookies.Session);
   const headers = req.headers;
   const headerToken = headers.get('authorization')?.replace('Bearer ', '');
@@ -27,7 +27,10 @@ export const authenticateRequest = async (req: NextRequest, opts: WithAuthOption
     secretKey: opts.secretKey || SECRET_KEY,
     frontendApi: opts.frontendApi || FRONTEND_API,
     publishableKey: opts.publishableKey || PUBLISHABLE_KEY,
-    signInUrl: opts.signInUrl || SIGN_IN_URL,
+    isSatellite,
+    domain,
+    signInUrl,
+    proxyUrl,
     cookieToken,
     headerToken,
     clientUat: getCookie(req, constants.Cookies.ClientUat),
